Добавляет модульность кода
